[DF-11566] more warning fixes
Former-commit-id: 5a4907eac1b4e30cf52a7814b7ddff111a14c265

+++ Sources/Internal/Platform/TemplateWin32/CorePlatformWin32.cpp
@@ -911,7 +911,7 @@ LRESULT CALLBACK CoreWin32Platform::WndProc(HWND hWnd, UINT message, WPARAM wPar
 
     // win32 app don't have ui-scaling option,
     // so hard-code default
-    float64 uiScale = 1.0;
+    float32 uiScale = 1.0;
 
     if (IsMouseInputEvent(message, GetMessageExtraInfo()))
     {
